:lock: Handle Auth in Controller to what User can do

+++ src/main/java/br/com/metaway/petshop/controllers/PetController.java
@@ -7,6 +7,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,8 +31,20 @@ public class PetController implements PetDocumentation {
 	@Autowired
 	private PetService service;
 	
+	@GetMapping("/daba")
+	public GrantedAuthority getDabadaou(Authentication auth) {
+		GrantedAuthority role = auth.getAuthorities().iterator().next();
+		System.out.println(role);
+		return role;
+	}
+	
 	@GetMapping
 	public ResponseEntity<Page<PetData>> index(Pageable pageable) {
+		if (service.containRole("USER")) {
+			Page<PetData> petsOfUser = service.getPetsByClientCpf(SecurityContextHolder.getContext().getAuthentication().getName(), pageable);
+			return ResponseEntity.ok(petsOfUser);
+		}
+
 		Page<PetData> pets = service.getAll(pageable);
 		return ResponseEntity.ok(pets);
 	}
@@ -42,11 +57,24 @@ public class PetController implements PetDocumentation {
 			return ResponseEntity.notFound().build();
 		}
 		
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		
+		if (service.containRole("USER") && pet.clientCpf() != username) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+		
 		return ResponseEntity.ok(pet);
 	}
 	
 	@PostMapping
     public ResponseEntity<PetData> store(@RequestBody Pet pet) {
+		PetData search = service.getById(pet.getId());
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+		if (service.containRole("USER") && search.clientCpf() != username) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+		
 		PetData newPet = service.create(pet);
 		
 		if (newPet == null) {
@@ -69,6 +97,10 @@ public class PetController implements PetDocumentation {
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> destroy(@PathVariable BigInteger id) {
+		if (service.containRole("USER")) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+		
 		Pet pet = service.delete(id);
 
 		if (pet == null) {
